Aircraft classes are added

+++ src/main/java/com/FourWings/atcSystem/model/airport/AirportsRepository.java
@@ -2,6 +2,6 @@ package com.FourWings.atcSystem.model.airport;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AirportsRepository extends JpaRepository<Airports, Integer> {
+public interface AirportsRepository extends JpaRepository<Airports, Long> {
     Airports findTopByOrderByAirportIdDesc();
 }
